+++ moor/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - It's now possible to change the name of a class generated for moor-file queries. See
   [the documentation](https://moor.simonbinder.eu/docs/using-sql/moor_files/#result-class-names) for details.
+- Added the `multiLine`, `caseSensitive`, `unicode` and `doAll` flags to the `regex` sql extension
+  method.
+  They require the `moor_ffi` version `0.7.0` or later.
 
 ## 3.1.0